requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
entsoe-py==0.6.9
anthropic==0.25.1
pydantic==2.7.1
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.7
